requirements.txt
# Principales
pandas>=2.0.0
numpy==1.26.4
sentence-transformers==3.0.1
scikit-learn>=1.3.0
openai>=1.12.0
python-dotenv>=0.21.0

# Frameworks de ML/NLP
torch
transformers>=4.38.0

# Otros
tabulate>=0.9.0
huggingface_hub>=0.20.0
